+++ public/filter.js
@@ -1,5 +1,0 @@
-document.querySelectorAll(".filter-form input[type=checkbox]").forEach((checkbox) => {
-  checkbox.addEventListener("change", () => {
-    checkbox.closest("form").submit();
-  });
-});