+++ src/middleware/rateLimiter.ts
@@ -1,7 +1,9 @@
 import { NextFunction, Request, Response } from "express";
-import { RateLimiterMemory } from "rate-limiter-flexible";
+import { RateLimiterPostgres } from "rate-limiter-flexible";
+import { pool } from "../db/db.js";
 
-const rateLimiter = new RateLimiterMemory({
+const rateLimiter = new RateLimiterPostgres({
+  storeClient: pool,
   keyPrefix: "middleware",
   points: 10, // requests
   duration: 1, // seconds